Move the spotify services to another package
